+++ .eslintrc.json
@@ -24,6 +24,7 @@
     "react/display-name": "off",
     "@typescript-eslint/no-explicit-any": "off",
     "@typescript-eslint/no-var-requires": 0,
+    "prettier/prettier": ["error", { "endOfLine": "auto" }],
     "import/order": [
       "error",
       {
