+++ src/app/(groups)/login/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+function Login() {
+  return <div> LOGIN </div>;
+}
+
+export default Login;
