+++ src/app/styles/globals.css
@@ -1,0 +1,9 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+body {
+  max-width: 1440px;
+  margin: 0 auto;
+  width: 100%;
+}
